requirements.txt
#azure_storage==0.37.0
fastapi==0.123.0
langchain_chroma==1.0.0
langchain_community==0.4.1
langchain_core==1.1.0
langchain_openai==1.1.0
langchain_text_splitters==1.0.0
langgraph==1.0.4
openai==2.8.1
pyaudio==0.2.14
pydantic==2.12.5
python-dotenv==1.2.1
Requests==2.32.5
webrtcvad==2.0.10
